requirements.txt
fasthtml
requests
python-dotenv
gunicorn
uvicorn
